[weather_level_adj] Fix bug I introduced with copy/paste errors.
The history_info, today_info, and forecast_info functions are not members of
the class and therefore don't have the self object to add the status.

Now we pass the main weather level checker class to the helper functions so it
can add status messages including when data was parsed incorrectly and even
use the pretty date from the JSON data to alert users on the status page.